Refactor: 혈당 카드 페이지 분리
